+++ src/components/content/DemoComponents/VideoTextSideBySide.stories.tsx
@@ -17,31 +17,21 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import React from 'react'
-import { SharedThemeProvider } from '../src/components/basic/SharedThemeProvider'
-import { SharedCssBaseline } from '../src/components/basic/SharedCssBaseline'
-import { I18nextProvider } from 'react-i18next'
-import i18n from '../src/i18next'
+import { VideoTextSideBySide as Component } from './VideoTextSideBySide'
+import type { Meta, StoryObj } from '@storybook/react'
 
-const withI18next = (Story) => {
-  return (
-    <SharedThemeProvider>
-      <SharedCssBaseline />
-      <I18nextProvider i18n={i18n}>
-        <Story />
-      </I18nextProvider>
-    </SharedThemeProvider>
-  );
-};
+const meta: Meta<typeof Component> = {
+  component: Component,
+}
 
-export const decorators = [withI18next];
+export default meta
 
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
+export const Default: StoryObj<typeof Component> = {
+  args : {
+  provider: {
+    title: "Example title",
+    description: "Example description for the component",
+    videoUrl: "https://eclipse-tractusx.github.io/video/traceability-video-9min-UT.mp4", 
+    }
+  }
 }