+++ src/components/content/DemoComponents/components/CardWithoutImage.tsx
@@ -17,31 +17,28 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import React from 'react'
-import { SharedThemeProvider } from '../src/components/basic/SharedThemeProvider'
-import { SharedCssBaseline } from '../src/components/basic/SharedCssBaseline'
-import { I18nextProvider } from 'react-i18next'
-import i18n from '../src/i18next'
+import { type CardDetailsProps } from '../ContentComponentsTypes'
+import { CardWithText } from './CardWithText'
+import { Box } from '@mui/material'
+import '../ContentComponents.scss'
 
-const withI18next = (Story) => {
+export const CardWithoutImage = ({
+  detail,
+  grid = 3,
+}: {
+  detail: CardDetailsProps
+  grid: number
+}) => {
   return (
-    <SharedThemeProvider>
-      <SharedCssBaseline />
-      <I18nextProvider i18n={i18n}>
-        <Story />
-      </I18nextProvider>
-    </SharedThemeProvider>
-  );
-};
-
-export const decorators = [withI18next];
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
+    <Box
+      key={detail.id}
+      className={'cardContainer'}
+      sx={{
+        backgroundColor: detail.backgroundColor,
+        width: `${100 / grid}%`,
+      }}
+    >
+      <CardWithText card={detail} isImage={false} />
+    </Box>
+  )
 }